backend/comparison_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
beautifulsoup4==4.12.2
rapidfuzz==3.5.2
python-dotenv==1.0.0
pydantic==2.4.2
aiohttp==3.9.3
requests==2.31.0
lxml==4.9.3
python-multipart==0.0.6
tenacity==8.2.3 